+++ src/scheduler.ml
@@ -2,38 +2,161 @@ open! Stdune
 open Import
 open Fiber.O
 
-type status_line_config =
-  { message   : string option
-  ; show_jobs : bool
-  }
-
-type running_job =
+type job =
   { pid  : int
   ; ivar : Unix.process_status Fiber.Ivar.t
   }
 
-module Watch : sig
-  (** Runs forever, communicating events to channel when changes are
-      detected. *)
-  val wait_to_chan : Path.t list Event.channel -> 'a
+(** The event queue *)
+module Event : sig
+  type t =
+    | Files_changed
+    | Job_completed of job * Unix.process_status
 
-  (** Compares a batch of changed files to ignored files list.
+  (** Return the next event. File changes event are always flattened
+      and returned first. *)
+  val next : unit -> t
 
-      This should be only called from the main thread. *)
-  val should_trigger_rebuild : Path.t list -> bool
+  (** Ignore the ne next file change event about this file. *)
+  val ignore_next_file_change_event : Path.t -> unit
 
-  (** Ignore the next event about this file. *)
-  val ignore_file : Path.t -> unit
+  (** Register a job that has just started *)
+  val register_job : job -> unit
+
+  (** Number of jobs for which the status hasn't been reported yet *)
+  val pending_jobs : unit -> int
+
+  (** Must be called from another thread than the main one. *)
+  val send_files_changed : Path.t list -> unit
+  type 'a process =
+    | Job : job -> OS.Type.windows process
+    | Pid : int -> OS.Type.unix process
+  val send_process_completed : OS.t process -> Unix.process_status -> unit
 end = struct
-  let ignored_files = String.Table.create 64
+  type t =
+    | Files_changed
+    | Job_completed of job * Unix.process_status
 
-  let ignore_file path =
+  type 'a process =
+    | Job : job -> OS.Type.windows process
+    | Pid : int -> OS.Type.unix process
+
+  let processes_completed : (OS.t process * Unix.process_status) Queue.t =
+    Queue.create ()
+  let files_changed = ref []
+  let mutex = Mutex.create ()
+  let cond = Condition.create ()
+
+  let ignored_files = String.Table.create 64
+  type _ jobs =
+    | Windows : int ref -> OS.Type.windows jobs
+    | Unix : (int, job) Hashtbl.t -> OS.Type.unix jobs
+  let jobs =
+    let f : type t. t OS.Type.t -> t jobs = function
+      | OS.Type.Windows -> Windows (ref 0)
+      | OS.Type.Unix -> Unix (Hashtbl.create 128)
+    in
+    f OS.t
+
+  let register_job job =
+    match jobs with
+    | Windows n -> incr n
+    | Unix tbl ->
+      assert (not (Hashtbl.mem tbl job.pid));
+      Hashtbl.add tbl job.pid job
+
+  let pending_jobs () =
+    match jobs with
+    | Windows n -> !n
+    | Unix tbl -> Hashtbl.length tbl
+
+  let ignore_next_file_change_event path =
     assert (Path.is_in_source_tree path);
     String.Table.replace
       ignored_files
       ~key:(Path.to_absolute_filename path)
       ~data:()
 
+  let available () =
+    not (List.is_empty !files_changed && Queue.is_empty processes_completed)
+
+  let event_of_process : type a. a jobs -> a process -> Unix.process_status -> t
+    = fun jobs process status ->
+      let job =
+        match jobs, process with
+        | Windows n, Job job ->
+          decr n;
+          job
+        | Unix tbl, Pid pid ->
+          match Hashtbl.find tbl pid with
+          | Some job ->
+            Hashtbl.remove tbl pid;
+            job
+          | None ->
+            Mutex.unlock mutex;
+            die "File watcher died: %s"
+              (match status with
+               | WEXITED n -> sprintf "exited with code %d" n
+               | WSIGNALED n -> sprintf "got signal %s" (Utils.signal_name n)
+               | WSTOPPED _ -> assert false)
+      in
+      Job_completed (job, status)
+
+  let next () =
+    Mutex.lock mutex;
+    let rec loop () =
+      if not (available ()) then Condition.wait cond mutex;
+      match !files_changed with
+      | [] ->
+        let (process, status) = Queue.pop processes_completed in
+        event_of_process jobs process status
+      | fns ->
+        files_changed := [];
+        let only_ignored_files =
+          List.fold_left fns ~init:true ~f:(fun acc fn ->
+            let fn = Path.to_absolute_filename fn in
+            if String.Table.mem ignored_files fn then begin
+              (* only use ignored record once *)
+              String.Table.remove ignored_files fn;
+              acc
+            end else
+              false)
+        in
+        if only_ignored_files then
+          loop ()
+        else
+          Files_changed
+    in
+    let ev = loop () in
+    Mutex.unlock mutex;
+    ev
+
+  let send_files_changed files =
+    Mutex.lock mutex;
+    let avail = available () in
+    files_changed := List.rev_append files !files_changed;
+    if not avail then Condition.signal cond;
+    Mutex.unlock mutex
+
+  let send_process_completed process status =
+    Mutex.lock mutex;
+    let avail = available () in
+    Queue.push (process, status) processes_completed;
+    if not avail then Condition.signal cond;
+    Mutex.unlock mutex
+end
+
+let ignore_for_watch = Event.ignore_next_file_change_event
+
+type status_line_config =
+  { message   : string option
+  ; show_jobs : bool
+  }
+
+module File_watcher : sig
+  (** Initializes the file watcher. *)
+  val init : unit -> unit
+end = struct
   let command =
     lazy (
       let excludes = [ {|/_build|}
@@ -104,8 +227,8 @@ end = struct
       ~len:!buffer_size;
     List.rev !lines
 
-  let wait_to_chan chan =
-    let new_events = ref [] in
+  let init = lazy(
+    let files_changed = ref [] in
     let event_mtx = Mutex.create () in
     let event_cv = Condition.create () in
 
@@ -130,7 +253,7 @@ end = struct
       let rec loop () =
         let lines = List.map (read_lines r) ~f:Path.of_string in
         Mutex.lock event_mtx;
-        new_events := lines :: !new_events;
+        files_changed := List.rev_append lines !files_changed;
         Condition.signal event_cv;
         Mutex.unlock event_mtx;
         loop ()
@@ -140,89 +263,70 @@ end = struct
 
     let rec buffer_thread () =
       Mutex.lock event_mtx;
-      if List.is_empty !new_events then
+      if List.is_empty !files_changed then
         Condition.wait event_cv event_mtx;
-      let events_to_send = !new_events in
-      new_events := [];
+      let files = !files_changed in
+      files_changed := [];
       Mutex.unlock event_mtx;
-      List.iter events_to_send ~f:(fun batch ->
-        Event.(sync (send chan batch)));
+      Event.send_files_changed files;
       Thread.delay buffering_time;
       buffer_thread ()
     in
 
-    ignore (Thread.create worker_thread ());
-    buffer_thread ()
+    ignore (Thread.create worker_thread () : Thread.t);
+    ignore (Thread.create buffer_thread () : Thread.t))
 
-  let should_trigger_rebuild fns =
-    List.filter fns ~f:(fun fn ->
-      let fn = Path.to_absolute_filename fn in
-      if String.Table.mem ignored_files fn then begin
-        (* only use ignored record once *)
-        String.Table.remove ignored_files fn;
-        false
-      end
-      else true)
-    |> List.is_empty
-    |> not
+  let init () = Lazy.force init
 end
 
-let ignore_for_watch = Watch.ignore_file
+module Process_watcher : sig
+  (** Initialize the process watcher thread. *)
+  val init : unit -> unit
 
-module Running_jobs : sig
-  val add : running_job -> unit
-  val count : unit -> int
+  (** Register a new running job *)
+  val register_job : job -> unit
 
-  (** Runs forever, communicating events to channel when children terminate. *)
-  val wait_to_chan : (int * Unix.process_status) Event.channel -> 'a
-
-  (** Removes a process previously returned via channel from the table. *)
-  val resolve_and_remove_job : int -> running_job
-
-  (** Send the following signal to all running jobs *)
+  (** Send the following signal to all running processes *)
   val killall : int -> unit
 end = struct
-  let all = Hashtbl.create 128
-
-  let running_pids = ref Int.Set.empty
-  let running_pids_mtx = Mutex.create ()
+  let jobs = Hashtbl.create 128
+  let jobs_mtx = Mutex.create ()
   let something_is_running_cv = Condition.create ()
 
-  let add job =
-    Hashtbl.add all job.pid job;
-    Mutex.lock running_pids_mtx;
-    running_pids := Int.Set.add !running_pids job.pid;
-    if Int.Set.cardinal !running_pids = 1 then
-      Condition.signal something_is_running_cv;
-    Mutex.unlock running_pids_mtx
+  let register_job job =
+    Event.register_job job;
+    Mutex.lock jobs_mtx;
+    let is_empty = Hashtbl.length jobs = 0 in
+    assert (not (Hashtbl.mem jobs job.pid));
+    Hashtbl.add jobs job.pid job;
+    if is_empty then Condition.signal something_is_running_cv;
+    Mutex.unlock jobs_mtx
 
-  let resolve_and_remove_job pid =
-    let job =
-      match Hashtbl.find all pid with
-      | Some job -> job
-      | None -> assert false
-    in
-    Hashtbl.remove all pid;
-    job
+  let killall signal =
+    Mutex.lock jobs_mtx;
+    Hashtbl.iter jobs ~f:(fun ~key:pid ~data:_ ->
+      try Unix.kill pid signal with _ -> ());
+    Mutex.unlock jobs_mtx
 
-  exception Finished of int * Unix.process_status
+  exception Finished of job * Unix.process_status
 
   let wait_nonblocking_win32 () =
-    Mutex.lock running_pids_mtx;
+    Mutex.lock jobs_mtx;
     match
-      Int.Set.iter !running_pids ~f:(fun pid ->
-        let pid, status = Unix.waitpid [WNOHANG] pid in
+      Hashtbl.iter jobs ~f:(fun ~key:_ ~data:job ->
+        let pid, status = Unix.waitpid [WNOHANG] job.pid in
         if pid <> 0 then
-          raise_notrace (Finished (pid, status)));
+          raise_notrace (Finished (job, status)))
     with
     | () ->
-      Mutex.unlock running_pids_mtx;
+      Mutex.unlock jobs_mtx;
       None
-    | exception (Finished (pid, status)) ->
-      Mutex.unlock running_pids_mtx;
-      Some (pid, status)
+    | exception (Finished (job, status)) ->
+      Hashtbl.remove jobs job.pid;
+      Mutex.unlock jobs_mtx;
+      Some (Event.Job job, status)
     | exception exn ->
-      Mutex.unlock running_pids_mtx;
+      Mutex.unlock jobs_mtx;
       reraise exn
 
   let rec wait_win32 () =
@@ -233,32 +337,34 @@ end = struct
     | Some x -> x
 
   let wait_unix () =
-    Unix.wait ()
+    let pid, status = Unix.wait () in
+    (Event.Pid pid, status)
 
   let wait =
-    if Sys.win32 then
-      wait_win32
-    else
-      wait_unix
+    let f : type t. t OS.Type.t -> unit -> t Event.process * Unix.process_status
+      = function
+        | OS.Type.Windows -> wait_win32
+        | OS.Type.Unix -> wait_unix
+    in
+    f OS.t
 
-  let rec wait_to_chan chan =
-    if Int.Set.is_empty !running_pids then
-      Condition.wait something_is_running_cv running_pids_mtx;
-    Mutex.unlock running_pids_mtx;
-    let pid, status = wait () in
-    Event.(sync (send chan (pid, status)));
-    Mutex.lock running_pids_mtx;
-    running_pids := Int.Set.remove !running_pids pid;
-    wait_to_chan chan
+  let run () =
+    Mutex.lock jobs_mtx;
+    while true do
+      if Hashtbl.length jobs = 0 then
+        Condition.wait something_is_running_cv jobs_mtx;
+      Mutex.unlock jobs_mtx;
+      let process, status = wait () in
+      Event.send_process_completed process status;
+      Mutex.lock jobs_mtx;
+      match process with
+      | Event.Job _ -> ()
+      | Event.Pid pid -> Hashtbl.remove jobs pid
+    done
 
-  let wait_to_chan chan =
-    Mutex.lock running_pids_mtx;
-    wait_to_chan chan
+  let init = lazy (ignore (Thread.create run () : Thread.t))
 
-  let count () = Hashtbl.length all
-
-  let killall signal =
-    Hashtbl.iter all ~f:(fun ~key:pid ~data:_ -> Unix.kill pid signal)
+  let init () = Lazy.force init
 end
 
 type t =
@@ -306,7 +412,7 @@ let update_status_line t =
     | { message = Some status_line; show_jobs } ->
       let status_line =
         if show_jobs then
-          sprintf "%s (jobs: %u)" status_line (Running_jobs.count ())
+          sprintf "%s (jobs: %u)" status_line (Event.pending_jobs ())
         else
           status_line
       in
@@ -327,7 +433,7 @@ let set_concurrency n =
 
 let wait_for_available_job () =
   Fiber.Var.get_exn t_var >>= fun t ->
-  if Running_jobs.count () < t.concurrency then
+  if Event.pending_jobs () < t.concurrency then
     Fiber.return t
   else begin
     let ivar = Fiber.Ivar.create () in
@@ -337,12 +443,12 @@ let wait_for_available_job () =
 
 let wait_for_process pid =
   let ivar = Fiber.Ivar.create () in
-  Running_jobs.add { pid; ivar };
+  Process_watcher.register_job { pid; ivar };
   Fiber.Ivar.read ivar
 
 let rec restart_waiting_for_available_job t =
   if Queue.is_empty t.waiting_for_available_job ||
-     Running_jobs.count () >= t.concurrency then
+     Event.pending_jobs () >= t.concurrency then
     Fiber.return ()
   else begin
     let ivar = Queue.pop t.waiting_for_available_job in
@@ -351,35 +457,11 @@ let rec restart_waiting_for_available_job t =
     restart_waiting_for_available_job t
   end
 
-let watch_channel = lazy (
-  if !Clflags.watch then begin
-    let chan = Event.new_channel () in
-    ignore (Thread.create Watch.wait_to_chan chan);
-    chan
-  end
-  else Event.new_channel ())
-
-let job_channel = lazy (
-  let chan = Event.new_channel () in
-  ignore (Thread.create Running_jobs.wait_to_chan chan);
-  chan)
-
-type event =
-  | Files_changed of Path.t list
-  | Job_completed of int * Unix.process_status
-
 let go_rec t =
-  let recv () = Event.(
-    choose [ wrap (receive (Lazy.force watch_channel))
-               (fun fns -> Files_changed fns)
-           ; wrap (receive (Lazy.force job_channel))
-               (fun (p, s) -> Job_completed (p, s))
-           ])
-  in
   let rec go_rec t =
     Fiber.yield ()
     >>= fun () ->
-    let count = Running_jobs.count () in
+    let count = Event.pending_jobs () in
     if count = 0 then begin
       hide_status_line t.status_line;
       flush stderr;
@@ -387,20 +469,16 @@ let go_rec t =
     end else begin
       update_status_line t;
       begin
-        match Event.sync (recv ()) with
-        | Job_completed (pid, status) ->
-          let job = Running_jobs.resolve_and_remove_job pid in
+        match Event.next () with
+        | Job_completed (job, status) ->
           Fiber.Ivar.fill job.ivar status
           >>= fun () ->
           restart_waiting_for_available_job t
           >>= fun () ->
           go_rec t
-        | Files_changed fns ->
-          if Watch.should_trigger_rebuild fns then begin
-            t.cur_build_canceled <- true;
-            Fiber.return ()
-          end
-          else go_rec t
+        | Files_changed ->
+          t.cur_build_canceled <- true;
+          Fiber.return ()
       end
     end
   in
@@ -410,6 +488,8 @@ let prepare ?(log=Log.no_log) ?(config=Config.default)
       ?(gen_status_line=fun () -> { message = None; show_jobs = false }) () =
   Log.infof log "Workspace root: %s"
     (Path.to_absolute_filename Path.root |> String.maybe_quoted);
+  if !Clflags.watch then File_watcher.init ();
+  Process_watcher.init ();
   let cwd = Sys.getcwd () in
   if cwd <> initial_cwd then
     Printf.eprintf "Entering directory '%s'\n%!"
@@ -465,18 +545,15 @@ let go ?log ?config ?gen_status_line fiber =
   run t (fun () -> fiber)
 
 let poll ?log ?config ~once ~finally ~canceled () =
-  let watch_chan = Lazy.force watch_channel in
   let t = prepare ?log ?config () in
   let once () =
     t.cur_build_canceled <- false;
     once ()
   in
-  let rec block_waiting_for_changes () =
-    let fns = Event.(sync (receive watch_chan)) in
-    if Watch.should_trigger_rebuild fns then
-      ()
-    else
-      block_waiting_for_changes ()
+  let block_waiting_for_changes () =
+    match Event.next () with
+    | Job_completed _ -> assert false
+    | Files_changed -> ()
   in
   let wait_success () =
     let old_generator = t.gen_status_line in
@@ -514,8 +591,7 @@ let poll ?log ?config ~once ~finally ~canceled () =
       wait_failure ()
       >>= fun _ ->
       main_loop ()
-    end
-    else begin
+    end else begin
       set_status_line_generator
         (fun () ->
            { message = Some "Had errors.\nKilling current build..."
@@ -523,10 +599,11 @@ let poll ?log ?config ~once ~finally ~canceled () =
            })
       >>= fun () ->
       Queue.clear t.waiting_for_available_job;
-      Running_jobs.killall Sys.sigkill;
-      while Running_jobs.count () > 0 do
-        let pid, _ = Event.sync (Event.receive (Lazy.force job_channel)) in
-        ignore (Running_jobs.resolve_and_remove_job pid : running_job);
+      Process_watcher.killall Sys.sigkill;
+      while Event.pending_jobs () > 0 do
+        match Event.next () with
+        | Files_changed -> ()
+        | Job_completed _ -> ()
       done;
       canceled ();
       main_loop ()