+++ src/stdune/stdune.ml
@@ -29,6 +29,7 @@ module Univ_map   = Univ_map
 module Loc        = Loc
 module Env        = Env
 module Proc       = Proc
+module OS         = OS
 
 external reraise : exn -> _ = "%reraise"